+++ test/integration/local/test_mnist.py
@@ -2,8 +2,7 @@ import os
 
 import numpy as np
 
-from test.utils import local_mode
-from test.utils.local_mode import request
+from test.utils import local_mode, test_utils
 
 mnist_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), '..', '..', 'resources', 'mnist')
 data_dir = os.path.join(mnist_path, 'data')
@@ -17,7 +16,7 @@ def test_chainer_mnist_single_machine(docker_image, opt_ml, use_gpu):
 
     files = ['model/model.npz', 'output/success', 'output/data/accuracy.png',
                                'output/data/cg.dot', 'output/data/log', 'output/data/loss.png']
-    _assert_files_exist(opt_ml, files)
+    test_utils.files_exist(opt_ml, files)
     assert not local_mode.file_exists(opt_ml, 'output/failure'), 'Failure happened'
     with local_mode.serve(os.path.join(mnist_path, customer_script), model_dir=None, image_name=docker_image,
                           opt_ml=opt_ml, use_gpu=use_gpu):
@@ -25,12 +24,12 @@ def test_chainer_mnist_single_machine(docker_image, opt_ml, use_gpu):
                        np.zeros((100, 28, 28), dtype='float32')]
         request_data = np.zeros((100, 784), dtype='float32')
         data_as_list = request_data.tolist()
-        _predict_and_assert_response_length(data_as_list, 'text/csv')
+        test_utils.predict_and_assert_response_length(data_as_list, 'text/csv')
         for array in test_arrays:
             # JSON and NPY can take multidimensional (n > 2) arrays
             data_as_list = array.tolist()
-            _predict_and_assert_response_length(data_as_list, 'application/json')
-            _predict_and_assert_response_length(request_data, 'application/x-npy')
+            test_utils.predict_and_assert_response_length(data_as_list, 'application/json')
+            test_utils.predict_and_assert_response_length(request_data, 'application/x-npy')
 
 
 def test_chainer_mnist_custom_loop(docker_image, opt_ml, use_gpu):
@@ -41,16 +40,16 @@ def test_chainer_mnist_custom_loop(docker_image, opt_ml, use_gpu):
                      source_dir=mnist_path, use_gpu=use_gpu)
 
     files = ['model/model.npz', 'output/success']
-    _assert_files_exist(opt_ml, files)
+    test_utils.files_exist(opt_ml, files)
     assert not local_mode.file_exists(opt_ml, 'output/failure'), 'Failure happened'
 
     with local_mode.serve(os.path.join(mnist_path, customer_script), model_dir=None, image_name=docker_image,
                           opt_ml=opt_ml):
         request_data = np.zeros((100, 784), dtype='float32')
         data_as_list = request_data.tolist()
-        _predict_and_assert_response_length(data_as_list, 'application/json')
-        _predict_and_assert_response_length(data_as_list, 'text/csv')
-        _predict_and_assert_response_length(request_data, 'application/x-npy')
+        test_utils.predict_and_assert_response_length(data_as_list, 'application/json')
+        test_utils.predict_and_assert_response_length(data_as_list, 'text/csv')
+        test_utils.predict_and_assert_response_length(request_data, 'application/x-npy')
 
 
 def test_chainer_mnist_distributed(docker_image, opt_ml, use_gpu):
@@ -69,23 +68,13 @@ def test_chainer_mnist_distributed(docker_image, opt_ml, use_gpu):
     files = ['model/model.npz', 'output/success', 'output/data/algo-1/accuracy.png',
              'output/data/algo-1/cg.dot', 'output/data/algo-1/log', 'output/data/algo-1/loss.png']
 
-    _assert_files_exist(opt_ml, files)
+    test_utils.files_exist(opt_ml, files)
     assert not local_mode.file_exists(opt_ml, 'output/failure'), 'Failure happened'
 
     with local_mode.serve(os.path.join(mnist_path, customer_script), model_dir=None, image_name=docker_image,
                           opt_ml=opt_ml):
         request_data = np.zeros((100, 784), dtype='float32')
         data_as_list = request_data.tolist()
-        _predict_and_assert_response_length(data_as_list, 'application/json')
-        _predict_and_assert_response_length(data_as_list, 'text/csv')
-        _predict_and_assert_response_length(request_data, 'application/x-npy')
-
-
-def _predict_and_assert_response_length(data, content_type):
-    predict_response = request(data, request_type=content_type)
-    assert len(predict_response) == len(data)
-
-
-def _assert_files_exist(opt_ml, files):
-    for file in files:
-        assert local_mode.file_exists(opt_ml, file), 'file {} was not created'.format(file)
+        test_utils.predict_and_assert_response_length(data_as_list, 'application/json')
+        test_utils.predict_and_assert_response_length(data_as_list, 'text/csv')
+        test_utils.predict_and_assert_response_length(request_data, 'application/x-npy')
